+++ compatibility-kit/javascript/features/minimal/minimal.ts
@@ -1,6 +1,6 @@
 import assert from 'assert'
 import { Given } from '@cucumber/fake-cucumber'
 
-Given('I have {int} cukes in my belly', function(cukeCount: number) {
+Given('I have {int} cukes in my belly', function (cukeCount: number) {
   assert(cukeCount)
 })