+++ gherkin/javascript/src/index.ts
@@ -1,4 +1,4 @@
-import Gherkin from './Gherkin'
+import GherkinStreams from './stream/GherkinStreams'
 import IGherkinOptions from './IGherkinOptions'
 import Query from './Query'
 import Dialect from './Dialect'
@@ -6,11 +6,15 @@ import Parser from './Parser'
 import AstBuilder from './AstBuilder'
 import TokenScanner from './TokenScanner'
 import compile from './pickles/compile'
+import DIALECTS from './gherkin-languages.json'
 
-export default Gherkin
+const dialects = DIALECTS as Readonly<{ [key: string]: Dialect }>
+
 export {
+  GherkinStreams,
   IGherkinOptions,
   Query,
+  dialects,
   Dialect,
   Parser,
   AstBuilder,